Virtualize the list of suggestions
There could be hundreds of errors for the big boards. This can affect the app performance
as we do not have any kind of pagination, chunk fetch, or other patters to retrieve data.
Thus we make sure that we do not render all items in the DOM tree, rather render only the
visible part of the suggestions. We use react-window to control the list
rendering