Create settings section
We want to add more features in the application.
Features like Sending Feedback, Fix all errors - require more
screen space and we do not have much
space in the iframe.
SO we add a new settings page where we
can hide all settings for the application.
Also there is now space for Fix all button.

In this commit we also fix the issue that
localStorage is not available in incognito mode
Now the application works but does not save the
language selection